+++ core-services/account-service/src/main/java/com/piggymetrics/account/AccountApplication.java
@@ -13,19 +13,23 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.cloud.security.oauth2.client.feign.OAuth2FeignRequestInterceptor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.client.DefaultOAuth2ClientContext;
 import org.springframework.security.oauth2.client.OAuth2RestTemplate;
 import org.springframework.security.oauth2.client.token.grant.client.ClientCredentialsResourceDetails;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableOAuth2Client;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
 import org.springframework.security.oauth2.provider.token.ResourceServerTokenServices;
 
 @SpringBootApplication
+@EnableResourceServer
 @EnableDiscoveryClient
 @EnableOAuth2Client
 @EnableFeignClients
 @EnableGlobalMethodSecurity(prePostEnabled = true)
 @EnableConfigurationProperties
-public class AccountApplication {
+public class AccountApplication extends ResourceServerConfigurerAdapter {
 
   @Autowired private ResourceServerProperties sso;
 
@@ -33,5 +37,34 @@ public class AccountApplication {
     SpringApplication.run(AccountApplication.class, args);
   }
 
+  @Bean
+  @ConfigurationProperties(prefix = "security.oauth2.client")
+  public ClientCredentialsResourceDetails clientCredentialsResourceDetails() {
+    return new ClientCredentialsResourceDetails();
+  }
 
+  @Bean
+  public RequestInterceptor oauth2FeignRequestInterceptor() {
+    return new OAuth2FeignRequestInterceptor(
+        new DefaultOAuth2ClientContext(), clientCredentialsResourceDetails());
+  }
+
+  @Bean
+  public OAuth2RestTemplate clientCredentialsRestTemplate() {
+    return new OAuth2RestTemplate(clientCredentialsResourceDetails());
+  }
+
+  @Bean
+  public ResourceServerTokenServices tokenServices() {
+    return new CustomUserInfoTokenServices(sso.getUserInfoUri(), sso.getClientId());
+  }
+
+  @Override
+  public void configure(HttpSecurity http) throws Exception {
+    http.authorizeRequests()
+        .antMatchers("/", "/demo", "/actuator/**")
+        .permitAll()
+        .anyRequest()
+        .authenticated();
+  }
 }
